Symlink iOS and MacOS to share native code
