+++ flutter/macos/sentry_flutter.podspec
@@ -13,8 +13,8 @@ Sentry SDK for Flutter with support to native through sentry-cocoa.
   s.source_files     = 'Classes/**/*'
   s.dependency 'Sentry', '~> 6.2.1'
   s.dependency 'FlutterMacOS'
-
   s.platform = :osx, '10.11'
+  
   s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
   s.swift_version = '5.0'
 end
