+++ flutter/pubspec.yaml
@@ -37,7 +37,7 @@ flutter:
       ios:
         pluginClass: SentryFlutterPlugin
       macos:
-        pluginClass: SentryFlutterPlugin
+        pluginClass: SwiftSentryFlutterPlugin
       web:
         pluginClass: SentryFlutterWeb
         fileName: sentry_flutter_web.dart
